support 128-bit Trace IDs when converting TraceIds and Kernels
